finally got back fully operable chat example
I finally have brought the project so far that it is now possible to send files
(images) through a websocket with sizes up to 6 MB (did not test more) and this
in addition of our new streaming api. This is really great. Furthermore I added
some harder tests for WebSocketStream.