rebuilding writting part of node-websockets
I mentioned that the WebSocketServer does not send the final frame on end of
piping. At least not to all clients (only to the uploading one...). In order to
fix this properly we must refactor a lot. For example we have the
WebSocketOutgoing we do not use. All the refactoring on the write part will be
happen in this pull request.